Added amazon clone's file
This project is a minimalist clone of Amazon, built solely with HTML and CSS within Visual Studio Code. 
By carefully crafting layout, styling, and interactive elements using pure HTML and CSS, this project underscores the importance of mastering the basics of front-end development. 
Dive into the codebase, explore the structure, and feel free to provide feedback or suggestions for improvement. 

Let's continue the journey of learning and refining our skills together!